MCP/requirements.txt
mcp-server>=0.1.0
pyserial>=3.5

